requirements.txt
numpy==1.24.3 # NOTE: Upgrading to 1.25.* might cause problems
opencv_python==4.8.*
mtcnn==0.1.1
pydub==0.25.1
setuptools==68.2.1
torch==2.0.1
nussl==1.1.9 # CFLAGS="-I/opt/homebrew/include" LDFLAGS="-L/opt/homebrew/lib" pip install nussl==1.1.9
tensorflow

# for linting
pylint
pylint-plugin-utils
autopep8
black

# performance metric
psutil==5.9.*

# below requirements are for audioseparator/tmp files
# flerken==0.2.4.post4
# gtts==2.3.1
# librosa==0.9.2
# moviepy==1.0.3
# soundfile==0.11.0

# testing
matplotlib==3.7.*
websockets==11.*
coverage==7.3.*
pytest-cov==4.1.*

# Django application
Django==4.2.*
celery==5.3.*
eventlet==0.33.*
daphne==4.0.*

# DRF
djangorestframework==3.14.*

# Django Channels
channels==4.0.*

# Install Redis server from https://github.com/microsoftarchive/redis/releases
